requirements.txt
pytest==8.0.2
nose2==0.14.1
pylint==3.1.0
requests==2.31.0
python-dotenv==1.0.1
# libs for fastapi
uvicorn==0.28.0

hvac~=2.1.0
pymsteams~=0.2.2
fastapi~=0.110.0
passlib~=1.7.4
pydantic~=2.6.4
